requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
alembic==1.12.1
alphashape==1.3.1
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
astroid==2.5.3
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.2
blinker==1.9.0
blis==1.3.3
brotli==1.2.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
Crawl4AI==0.7.4
crypto==1.4.1
cryptography==46.0.3
curl_cffi==0.12.0
cycler==0.10.0
cymem==2.0.13
dataclasses-json==0.6.7
ddgs==9.8.0
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
et_xmlfile==2.0.0
eval_type_backport==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.128.0
fastuuid==0.14.0
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.4
google-genai==1.28.0
google-generativeai==0.4.1
googleapis-common-protos==1.70.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.2.4
griffe==1.9.0
groq==0.30.0
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.16.0
h2==4.3.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
humanize==4.13.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==8.18.1
isort==5.8.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
JsonForm==0.0.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
JsonSir==0.0.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kconfiglib==14.1.0
kiwisolver==1.3.2
kombu==5.5.4
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.3.83
langchain-google-genai==1.0.1
langchain-openai==0.0.2
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lark==1.3.1
lazy-object-proxy==1.6.0
litellm==1.80.10
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.4.3
matplotlib-inline==0.1.7
mavsdk==3.5.0
mccabe==0.6.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
murmurhash==1.0.15
mypy_extensions==1.1.0
Naked==0.1.31
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
nodeenv==1.9.1
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.17.3
openai==2.14.0
openai-agents==0.2.4
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.11.0
packaging==25.0
pandas==1.3.4
parameterized==0.8.1
parso==0.8.4
patchright==1.56.0
pathlib==1.0.1
peewee==3.18.2
pexpect==4.9.0
phi==0.6.7
phidata==2.7.10
pillow==11.3.0
platformdirs==4.3.8
playwright==1.57.0
preshed==3.0.12
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylint==2.7.4
PyMuPDF==1.26.3
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==5.9.0
pyrsistent==0.17.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
Python-EasyConfig==0.1.7
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.0
rank-bm25==0.2.2
redis==5.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
Resource==0.2.1
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rtree==1.4.1
ruff==0.12.7
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==5.1.2
shapely==2.0.7
shellescape==3.8.1
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
six==1.15.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
socksio==1.0.0
soupsieve==2.7
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.2
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tavily-python==0.7.10
tenacity==8.5.0
tf-playwright-stealth==1.2.0
thinc==8.3.9
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.2.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
trimesh==4.10.1
typer==0.23.2
typer-slim==0.23.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.14.0
uvicorn==0.39.0
uvloop==0.22.1
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.12.1
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.65
zipp==3.23.0
zstandard==0.25.0
